+++ src/oihana/core/strings/randomKey.php
@@ -2,15 +2,13 @@
 
 namespace oihana\core\strings ;
 
-use oihana\enums\Char;
-
 /**
  * Generates a random key.
  * @param ?string $prefix The expression to format.
  * @param string $separator The separator characters/expression of the random key.
  * @return string The new random key.
  */
-function randomKey( ?string $prefix , string $separator = Char::UNDERLINE ) :string
+function randomKey( ?string $prefix , string $separator = '_' ) :string
 {
     return ( is_string( $prefix ) ? ( $prefix . $separator ) : '' ) . mt_rand() ;
 }